+++ facet-derive-emit/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add support for `Shape::type_tag`
+
 ## [0.27.7](https://github.com/facet-rs/facet/compare/facet-derive-emit-v0.27.6...facet-derive-emit-v0.27.7) - 2025-05-27
 
 ### Other