+++ facet-derive-parse/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add support for `Shape::type_tag`
+
 ## [0.27.0](https://github.com/facet-rs/facet/compare/facet-derive-parse-v0.26.1...facet-derive-parse-v0.27.0) - 2025-05-13
 
 ### Other